aded persistence with react-persist
