+++ src/index.js
@@ -1,16 +1,20 @@
+// import { PersitGate } from "redux-persist/integration/react";
+import { PersistGate } from "redux-persist/integration/react";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 import { render } from "react-dom";
 import React from "react";
-import "./index.scss";
 
-import store from './redux/store'
+import { store, persistor } from "./redux/store";
 import App from "./app";
+import "./index.scss";
 
 render(
   <Provider store={store}>
     <BrowserRouter>
-      <App />
+      <PersistGate persistor={persistor}>
+        <App />
+      </PersistGate>
     </BrowserRouter>
   </Provider>,
   document.getElementById("root")
